+++ src/main/java/org/betonquest/betonquest/quest/event/stage/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.event.Event Event} implementation of the stage event.
+ */
+package org.betonquest.betonquest.quest.event.stage;